requirements.txt
streamlit==1.35.0
joblib==1.3.2
plotly==5.19.0
Pillow==10.2.0
scikit-learn==1.3.2
